a la norme sans les header

+++ ft_strlcat.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-size_t  ft_strlcat(char *dst, const char *src, size_t dstsize)
+size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
 	size_t i;
 	size_t j;
@@ -8,7 +8,7 @@ size_t  ft_strlcat(char *dst, const char *src, size_t dstsize)
 
 	j = 0;
 	i = (size_t)ft_strlen(dst);
-	dstog = (size_t)ft_strlen(dst);                 
+	dstog = (size_t)ft_strlen(dst);             
 	while (src[j] && (j + dstog < dstsize - 1) && dstsize != 0)
 	{
 		dst[i] = src[j];
@@ -17,7 +17,7 @@ size_t  ft_strlcat(char *dst, const char *src, size_t dstsize)
 	}
 	dst[i] = '\0';
 	if (dstog >= dstsize)
-		return ((unsigned)ft_strlen(src) + dstsize);
+		return (ft_strlen(src) + dstsize);
 	else
-		return ((unsigned)ft_strlen(src) + dstog);
+		return (ft_strlen(src) + dstog);
 }
